* changed type of parameters to functions from float to double    - calling code was passing doubles but was being casted to float, but there      was not real need for function to be using floats

+++ calculation/utility/UtilityFunctions.cpp
@@ -28,19 +28,19 @@ double CalcLogLikelihood(count_t n, measure_t u, count_t N, measure_t U) {
 //
 //		Latitude and Longitude values should be float, ie, real numbers
 //		with ranges [-90,90] and [-180,180] (in degrees) respectively.
-void ConvertFromLatLong(float Latitude, float Longitude, std::vector<double>& vCoordinates) {
-  float RADIUS = 6367; // Constant; radius of earth in km)
+void ConvertFromLatLong(double Latitude, double Longitude, std::vector<double>& vCoordinates) {
+  double RADIUS = 6367; // Constant; radius of earth in km)
 
-  vCoordinates[0] = (double)(RADIUS * cos(Longitude*PI/180.0) * cos(Latitude*PI/180.0)); // x coordinate
-  vCoordinates[1] = (double)(RADIUS * sin(Longitude*PI/180.0) * cos(Latitude*PI/180.0)); // y coordinate
-  vCoordinates[2] = (double)(RADIUS * sin(Latitude*PI/180.0));														// z coordinate
+  vCoordinates[0] = (RADIUS * cos(Longitude*PI/180.0) * cos(Latitude*PI/180.0)); // x coordinate
+  vCoordinates[1] = (RADIUS * sin(Longitude*PI/180.0) * cos(Latitude*PI/180.0)); // y coordinate
+  vCoordinates[2] = (RADIUS * sin(Latitude*PI/180.0));														// z coordinate
 }
 
-void ConvertToLatLong(float* Latitude, float* Longitude, double* pCoords) {
-  float RADIUS = 6367; // Constant; radius of earth in km)
+void ConvertToLatLong(double* Latitude, double* Longitude, double* pCoords) {
+  double RADIUS = 6367; // Constant; radius of earth in km)
 
   if (pCoords[0] != 0) {
-    *Longitude = (float)(atan(pCoords[1] / pCoords[0]) * 180.0 / PI);
+    *Longitude = (atan(pCoords[1] / pCoords[0]) * 180.0 / PI);
     if (pCoords[0] < 0 && pCoords[1] > 0)
       *Longitude += 180.0;
     else if (pCoords[0] < 0 && pCoords[1] < 0)
@@ -53,8 +53,8 @@ void ConvertToLatLong(float* Latitude, float* Longitude, double* pCoords) {
   else if (pCoords[1] == 0)
     *Longitude = 0.0;
 
-  float tmp = (float)sqrt((pCoords[0]*pCoords[0] + pCoords[1]*pCoords[1])/(RADIUS*RADIUS));
-  *Latitude = (float)((pCoords[2] >= 0 ? (1.0) : (-1.0)) * acos(tmp) * 180.0 / PI);
+  double tmp = sqrt((pCoords[0]*pCoords[0] + pCoords[1]*pCoords[1])/(RADIUS*RADIUS));
+  *Latitude = ((pCoords[2] >= 0 ? (1.0) : (-1.0)) * acos(tmp) * 180.0 / PI);
 }
 
 void DisplayVersion(FILE* fp=stdout, int nPos=0)
